requirements.txt
# requirements.txt

MouseInfo==0.1.3
numpy==1.24.4
opencv-python==4.9.0.80
pillow==10.2.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySide6==6.6.2
PySide6-Addons==6.6.2
PySide6-Essentials==6.6.2
pytweening==1.2.0
shiboken6==6.6.2
